disabled at everyone

+++ onMessage.js
@@ -30,10 +30,20 @@ const onMessage = (client, prefix) => {
       }
       ////
       if (command === 'say') {
-        const sayMessage = args.join(' ');
-        msg.delete().catch(o3o => {});
-        msg.channel.send(sayMessage);
-        logOutput(msg, command, args);
+        ////
+        if (
+          command.includes('@everyone') ||
+          command.includes('@here') ||
+          args.includes('@everyone') ||
+          command.includes('@here')
+        ) {
+          msg.channel.send('no');
+        } else {
+          const sayMessage = args.join(' ');
+          msg.delete().catch(o3o => {});
+          msg.channel.send(sayMessage);
+          logOutput(msg, command, args);
+        }
       }
       ////
       if (command === 'delete') {
